+++ onnxscript/_internal/utils.py
@@ -5,16 +5,13 @@
 from __future__ import annotations
 
 import numbers
-from typing import Any, Iterable, Optional, Sequence
+from typing import Optional, Sequence
 
 import numpy as np
 import onnx
 import onnx.helper
-from onnx import FunctionProto, ModelProto, TensorProto, ValueInfoProto
 
 from onnxscript import tensor
-
-# pylint: enable=unused-import, ungrouped-imports
 
 
 def external_tensor(
@@ -26,7 +23,7 @@ def external_tensor(
     length: Optional[int] = None,
     checksum: Optional[str] = None,
     basepath: Optional[str] = None,
-) -> TensorProto:
+) -> onnx.TensorProto:
     """Create a TensorProto referencing externally stored tensor-data.
 
     Args:
@@ -44,11 +41,11 @@ def external_tensor(
 
     See https://github.com/onnx/onnx/blob/main/docs/ExternalData.md for more details.
     """
-    tensor_proto = TensorProto()
+    tensor_proto = onnx.TensorProto()
     tensor_proto.name = name
     tensor_proto.data_type = data_type
     tensor_proto.dims.extend(dims)
-    tensor_proto.data_location = TensorProto.EXTERNAL
+    tensor_proto.data_location = onnx.TensorProto.EXTERNAL
 
     def add(k, v):
         entry = tensor_proto.external_data.add()
@@ -74,9 +71,9 @@ def value_to_type_proto(val):
         shape = val.shape
         return onnx.helper.make_tensor_type_proto(elem_type, shape)
     if isinstance(val, int):
-        return onnx.helper.make_tensor_type_proto(TensorProto.INT32, [])
+        return onnx.helper.make_tensor_type_proto(onnx.TensorProto.INT32, [])
     if isinstance(val, (float, np.float32)):
-        return onnx.helper.make_tensor_type_proto(TensorProto.FLOAT, [])
+        return onnx.helper.make_tensor_type_proto(onnx.TensorProto.FLOAT, [])
     if isinstance(val, list):
         if len(val) > 0:
             return onnx.helper.make_sequence_type_proto(value_to_type_proto(val[0]))
@@ -84,7 +81,7 @@ def value_to_type_proto(val):
         # Should be using a typed-value instead.
         # Treated as a sequence of tensors of float-type.
         return onnx.helper.make_sequence_type_proto(
-            onnx.helper.make_tensor_type_proto(TensorProto.FLOAT, None)
+            onnx.helper.make_tensor_type_proto(onnx.TensorProto.FLOAT, None)
         )
     if isinstance(val, numbers.Number):
         nparray = np.array(val)
@@ -102,50 +99,3 @@ def values_to_value_infos(name_values):
         for (name, val) in name_values
         if val is not None
     ]
-
-
-def make_model_from_function_proto(
-    function_proto: FunctionProto,
-    function_opset_version: int,
-    input_value_infos: Sequence[ValueInfoProto],
-    output_value_infos: Sequence[ValueInfoProto],
-    **attrs: Any,
-) -> ModelProto:
-    """Creates a model containing a single call to a given
-    function with input and output value_infos, etc.
-
-    Args:
-        function_proto (FunctionProto): function proto
-            representing a single call
-        function_opset_version (int):  function_proto's version
-        input_value_infos (list of ValueInfoProto): function's input
-        output_value_infos (list of ValueInfoProto): function's output
-        **attrs (dict): the attributes of the node for the function
-
-    Returns:
-        ModelProto
-    """
-
-    input_names = [vi.name for vi in input_value_infos]
-    output_names = [vi.name for vi in output_value_infos]
-    node = onnx.helper.make_node(
-        function_proto.name,
-        input_names,
-        output_names,
-        domain=function_proto.domain,
-        **attrs,
-    )
-    graph = onnx.helper.make_graph([node], "node_graph", input_value_infos, output_value_infos)
-    model_proto_opset: Iterable[onnx.OperatorSetIdProto] = function_proto.opset_import
-    if all(o.domain != function_proto.domain for o in model_proto_opset):
-        model_proto_opset = [
-            *model_proto_opset,
-            onnx.helper.make_opsetid(function_proto.domain, function_opset_version),
-        ]
-    model = onnx.helper.make_model(
-        graph,
-        functions=[function_proto],
-        producer_name="onnxscript",
-        opset_imports=model_proto_opset,
-    )
-    return model
